+++ src/main/java/nextstep/ladder/ladder/LadderHeight.java
@@ -1,4 +1,4 @@
-package nextstep.ladder;
+package nextstep.ladder.ladder;
 
 import nextstep.ladder.exception.IllegalLadderHeightException;
