+++ src/main/java/nextstep/ladder/player/Player.java
@@ -1,4 +1,4 @@
-package nextstep.ladder;
+package nextstep.ladder.player;
 
 import nextstep.ladder.exception.IllegalPlayerNameSizeException;
