+++ src/main/java/nextstep/ladder/player/Players.java
@@ -1,4 +1,4 @@
-package nextstep.ladder;
+package nextstep.ladder.player;
 
 import nextstep.ladder.util.StringUtil;
