+++ src/test/java/nextstep/ladder/ladder/LadderTest.java
@@ -1,5 +1,7 @@
-package nextstep.ladder;
+package nextstep.ladder.ladder;
 
+import nextstep.ladder.ladder.Ladder;
+import nextstep.ladder.ladder.LadderHeight;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
