+++ src/test/java/nextstep/ladder/ladder/LineTest.java
@@ -1,5 +1,6 @@
-package nextstep.ladder;
+package nextstep.ladder.ladder;
 
+import nextstep.ladder.ladder.Line;
 import nextstep.ladder.strategy.RandomLadderStrategy;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
