+++ src/test/java/nextstep/ladder/player/PlayerTest.java
@@ -1,6 +1,7 @@
-package nextstep.ladder;
+package nextstep.ladder.player;
 
 import nextstep.ladder.exception.IllegalPlayerNameSizeException;
+import nextstep.ladder.player.Player;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
