+++ src/test/java/nextstep/ladder/player/PlayersTest.java
@@ -1,5 +1,6 @@
-package nextstep.ladder;
+package nextstep.ladder.player;
 
+import nextstep.ladder.player.Players;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
